# apps/api-gateway/openapi/openapi.yaml
openapi: 3.0.0
info:
  title: AP Calculus Bot API
  description: API Gateway for AP Calculus AB/BC tutoring bot with verified answers
  version: 0.1.0
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /payments/plans:
    get:
      tags:
        - payments
      description: Get available pricing plans
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: string
                    name:
                      type: string
                    description:
                      type: string
                    price:
                      type: number
                    currency:
                      type: string
                    interval:
                      type: string
                      enum:
                        - month
                        - year
                    features:
                      type: array
                      items:
                        type: string
                    role:
                      type: string
                  required:
                    - id
                    - name
                    - description
                    - price
                    - currency
                    - interval
                    - features
                    - role
                  additionalProperties: false
        "500":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: object
                    properties:
                      message:
                        type: string
                        description: Human readable error message
                      statusCode:
                        type: number
                        description: HTTP status code associated with the error
                      code:
                        type: string
                        description: Application-specific error code
                      validation:
                        description: Validation errors when applicable
                      details:
                        description: Additional error context
                    required:
                      - message
                      - statusCode
                    additionalProperties: false
                required:
                  - error
                additionalProperties: false
  /payments/checkout:
    post:
      tags:
        - payments
      description: Create Stripe checkout session
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                priceId:
                  type: string
              required:
                - priceId
              additionalProperties: false
        required: true
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  url:
                    type: string
                required:
                  - id
                  - url
                additionalProperties: false
        "400":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: object
                    properties:
                      message:
                        type: string
                        description: Human readable error message
                      statusCode:
                        type: number
                        description: HTTP status code associated with the error
                      code:
                        type: string
                        description: Application-specific error code
                      validation:
                        description: Validation errors when applicable
                      details:
                        description: Additional error context
                    required:
                      - message
                      - statusCode
                    additionalProperties: false
                required:
                  - error
                additionalProperties: false
        "500":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: object
                    properties:
                      message:
                        type: string
                        description: Human readable error message
                      statusCode:
                        type: number
                        description: HTTP status code associated with the error
                      code:
                        type: string
                        description: Application-specific error code
                      validation:
                        description: Validation errors when applicable
                      details:
                        description: Additional error context
                    required:
                      - message
                      - statusCode
                    additionalProperties: false
                required:
                  - error
                additionalProperties: false
  /payments/stripe/portal:
    post:
      tags:
        - payments
      description: Create Stripe billing portal session
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  url:
                    type: string
                required:
                  - url
                additionalProperties: false
        "400":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: object
                    properties:
                      message:
                        type: string
                        description: Human readable error message
                      statusCode:
                        type: number
                        description: HTTP status code associated with the error
                      code:
                        type: string
                        description: Application-specific error code
                      validation:
                        description: Validation errors when applicable
                      details:
                        description: Additional error context
                    required:
                      - message
                      - statusCode
                    additionalProperties: false
                required:
                  - error
                additionalProperties: false
        "500":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: object
                    properties:
                      message:
                        type: string
                        description: Human readable error message
                      statusCode:
                        type: number
                        description: HTTP status code associated with the error
                      code:
                        type: string
                        description: Application-specific error code
                      validation:
                        description: Validation errors when applicable
                      details:
                        description: Additional error context
                    required:
                      - message
                      - statusCode
                    additionalProperties: false
                required:
                  - error
                additionalProperties: false
...
servers:
  - url: http://localhost:3001
    description: Development server